Input subsystem timestamp implementation

+++ sys/include/sys/input.h
@@ -25,8 +25,8 @@
 
 // Input Event Structure (Linux compatible)
 typedef struct input_event {
-    uint32_t time_sec;
-    uint32_t time_usec;
+    uint64_t time_sec;
+    uint64_t time_usec;
     uint16_t type;
     uint16_t code;
     int32_t  value;
